[FIX] l10n_multilang: Getting translations of templates from records
Bug:

Before the the fix, when trying to get all the translations of templates from a
specific model:
-All the translations for the templates of this model were catched first -> TT
-All the records of this model for a specific company were catched afterwards -> RC

Then, the linked between the translations(TT) and the records(RC) were made in function
process_translations.

But in some cases, a few records were removed and then it raised an error because the number
of translations was different than the number of records.

Fix:

Thanks to the function create_record_with_xmlid in addons/account/models/chart_template.py, all the records
created from a template have an xml_id in ir.model.data with str(company.id)+'_'+template_xmlid.name

So with this way, it is possible to catch only the records created from a template.

opw:767095