[IMP] stock: sheduled date update with picking delivery type
When deliver all at once on a picking the expected date for it
was the minimum expected date between the stock move
for this picking. (same behavior as partial delivery)

This commit merge both min_date and max_date field in a single
field named shceduled_date. It replace min_date in views by
sheduled date that varies following picking move_type.