[FIX] portal: fix frontend chatter move
Commit 8e3111a233820529a606c2fa3dff16d40e3fcac0 moved the frontend chatter from a website-based module
website_mail to portal. Purpose was to allow its use in customer
portal independently from the website.

However some renaming done during the move was not properly propagated
resulting in the chatter not being loaded. This commit fixes that by
correctly using the new class node as well as widget name when
instantiating the widget.