[FIX] auth_signup: fix computation of default template user in settings
Indeed we store an integer that is the ID of the user to copy when
creating the portal user, not a boolean. We therefore have to evaluate
the result instead of comparing it to a boolean. This was introduced
at c9a53bd217e162c089f3170f672722b67c7e0d0b.