[FIX] mrp: update qty to produce does not remove the qty produced
When updating the quantity to produce on a MO, the quantity done
on finished/consumed moves that were not done was erased.

It happens due to write on stock move that always wanted to
unreserve the move and thus drop the move lines with qty_done.
This issue has been resolve in previous commit

This commit correct some mistakes in wizard methods in order
to correctly update open moves.