[FIX] website_portal_sale: fix override not giving correctly parameters to super
It does not have any impact in current stable as the inherited
method does not use it. However if another addon inherit it
it will not receive access_token depending on call chain. Let
us fix it.

This is a manual forward-port of ce9ed600ca617ffcda86dabb02df7ca30c340ea6.