[FIX] stock: move_view without suggest
The move line view when the option no suggest is checked doesn't work
because it hasen't been adapted during the refactoring of stock module.

So to avoid having differences between the view in no suggest and in
suggest, we use the same tree view that we reference. Because the views
shouldn't be different if we are either in suggest or no suggest